refactor(button, link): add missing base, use new stylesheets
Followup to #615
